- Add owner url column to owner table - Index both the above and url column of character image creators - Consolidate alias update commands into one and update character image creator command for new system
